+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Added:
 
-* Initial release with filesystem check
+* Initial release
+* filesystem check
+* http check
 
 [Unreleased]: https://github.com/julian7/sensu-base-checks